Fixed error handling in API calls
Previously many errors were being swallowed by Plaidster (anything that hit the catch block). This would not only hide the error from the application, but also fail to call the handler which breaks flow control when nesting API calls in completion handlers.

I've improved the PlaidError and PlaidsterError ErrorTypes to allow them to be easily converted to NSErrors. Then I improved the PlaidApi functions to convert exceptions to NSErrors and call their handlers on failures. Also I consolidated the JSONError and PlaidError types as well as made all error types public so they can be used in the enclosing application (for example to check error codes).

Eventually I'll add PlaidErrors for more Plaid error codes, but this lays the foundation for proper error handling.